requirements.txt
datasets
evaluate
transformer[sentencepiece]
accelerate
